requirements.txt
rdoclient
django-excel
django-mathfilters
django-compressor
django-channels